Split up votingMethodConfig for better readability
